+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Support Asymmetric Keys in KMS
 - Support for Blockchain API improvements
+- EXaCC ADBD Data Guard + ACL Support
 
 ## 4.10.0 (January 20, 2021)